chore(app): update 'studio' to use newly created 'core' lib
