+++ apps/studio/package.json
@@ -13,7 +13,7 @@
   },
   "dependencies": {
     "@biolnk/gamut": "*",
-    "@biolnk/utils": "*",
+    "@biolnk/core": "*",
     "@headlessui/react": "^1.4.2",
     "@hookform/resolvers": "^2.8.4",
     "@supabase/supabase-js": "^1.28.4",
@@ -29,7 +29,6 @@
     "vest": "^4.0.1"
   },
   "devDependencies": {
-    "@biolnk/config": "*",
     "@tailwindcss/forms": "^0.4.0",
     "@tailwindcss/typography": "^0.5.0",
     "@testing-library/jest-dom": "^5.15.1",
