+++ apps/studio/src/components/account/AccountSkeleton.tsx
@@ -1,7 +1,7 @@
 import Skeleton from "react-loading-skeleton";
 import { AccountLayout } from "~/components/layouts";
 import { Flex } from "@biolnk/gamut";
-import { FC } from "react";
+import type { FC } from "react";
 
 const AccountSkeleton: FC = () => {
   return (
