+++ apps/studio/src/components/common/Header/Header.tsx
@@ -3,9 +3,9 @@ import Logo from "~/assets/images/biolnk.png";
 import Menu from "./Menu";
 import Link from "../Link";
 import useUser from "~/utils/hooks/queries/useUser";
-import useClipboard from "~/utils/hooks/useClipboard";
 import getPageLink from "~/utils/misc/getPageLink";
 import { memo } from "react";
+import { useClipboard } from "@biolnk/core";
 import { BaseIcon, Container, Text, Copy, Button, Flex } from "@biolnk/gamut";
 
 import Styles from "./Header.module.css";
