+++ apps/studio/src/components/common/Link.tsx
@@ -1,6 +1,6 @@
 import React, { forwardRef } from "react";
 import NextLink, { LinkProps as NextLinkProps } from "next/link";
-import { ctl } from "@biolnk/utils";
+import { ctl } from "@biolnk/core";
 import { BaseIcon, ExternalLink } from "@biolnk/gamut";
 
 export interface LinkProps
