+++ apps/studio/src/components/dashboard/PageNavigation.tsx
@@ -8,8 +8,8 @@ import {
   BarChart2,
   Trello,
 } from "@biolnk/gamut";
+import { ctl } from "@biolnk/core";
 import { Routes } from "~/data/enums/routes";
-import { ctl } from "@biolnk/utils";
 
 const PageNavigation: React.FC = () => {
   const { pathname } = useRouter();
