+++ apps/studio/src/components/dashboard/PageSeoForm.tsx
@@ -4,7 +4,8 @@ import useUpdatePage from "~/utils/hooks/mutations/useUpdatePage";
 import { FC, memo } from "react";
 import { Button, Flex, Input, Textarea } from "@biolnk/gamut";
 import { PAGE_SEO_SCHEMA } from "~/data/validations";
-import { PageSeoDto, PageWithMetadata } from "~/types";
+import type { PageWithMetadata } from "@biolnk/core";
+import type { PageSeoDto } from "~/types";
 
 export interface PageSeoFormProps {
   page: PageWithMetadata;
