+++ apps/studio/src/components/dashboard/ThemeCard.tsx
@@ -1,8 +1,6 @@
-import CSSstring from "~/utils/CSSstring";
 import { ChangeEvent, FC } from "react";
 import { Flex, Input } from "@biolnk/gamut";
-import { ctl } from "@biolnk/utils";
-import { Theme } from "~/types";
+import { Theme, CSSstring, ctl } from "@biolnk/core";
 
 export interface ThemeCardProps extends Theme {
   userCurrentTheme: string;
