+++ apps/studio/src/components/dashboard/ThemeListSkeleton.tsx
@@ -1,4 +1,4 @@
-import { FC } from "react";
+import type { FC } from "react";
 
 const ThemeListSkeleton: FC = () => {
   return (
