+++ apps/studio/src/data/context/app.tsx
@@ -1,8 +1,5 @@
-import useDisclosure, {
-  UseDisclosureReturn,
-} from "~/utils/hooks/useDisclosure";
-import { makeContext } from "~/utils/makeContext";
-import { ReactNode } from "react";
+import { makeContext, useDisclosure, UseDisclosureReturn } from "@biolnk/core";
+import type { ReactNode } from "react";
 
 export type AppContextState = {
   addLinkDialog: UseDisclosureReturn;
