+++ apps/studio/src/data/enums/index.ts
@@ -1,3 +1,1 @@
-export * from "./db";
 export * from "./routes";
-export * from "./tables";
