--- apps/studio/src/data/enums/tables.ts
@@ -1,6 +1,0 @@
-export enum Tables {
-  USERS = "users",
-  PAGES = "pages",
-  LINKS = "links",
-  THEMES = "themes",
-}
