+++ apps/studio/src/lib/supabase/provider.tsx
@@ -14,9 +14,9 @@ import {
   getUserById,
   getUserByUsername,
 } from "~/services/supabase";
-import { makeContext } from "~/utils/makeContext";
+import { makeContext, User } from "@biolnk/core";
 import type { Session } from "@supabase/supabase-js";
-import type { SignInDto, SignUpDto, User, AuthUser } from "~/types";
+import type { SignInDto, SignUpDto, AuthUser } from "~/types";
 
 export type SupabaseContextState = {
   user: User;
