+++ apps/studio/src/types/index.ts
@@ -1,6 +1,2 @@
 export * from "./auth";
 export * from "./dtos";
-export * from "./user";
-export * from "./link";
-export * from "./page";
-export * from "./theme";
