+++ apps/studio/src/utils/HOC/withAuthCheck.tsx
@@ -1,5 +1,5 @@
 import useUser from "~/utils/hooks/queries/useUser";
-import useSafeLayoutEffect from "../hooks/useSafeLayoutEffect";
+import { useSafeLayoutEffect } from "@biolnk/core";
 import { ComponentType, useState } from "react";
 import { useRouter } from "next/router";
 import { Loading } from "@biolnk/gamut";
