+++ apps/studio/src/utils/hooks/mutations/useCreateNewLink.ts
@@ -1,8 +1,8 @@
 import { useMutation, useQueryClient } from "react-query";
 import { makeToast } from "@biolnk/gamut";
+import { LinkKind } from "@biolnk/core";
 import { sbClient } from "~/lib/supabase";
 import { createNewLink } from "~/services/supabase";
-import { LinkKind } from "~/data/enums";
 import type { CreateLinkDto, FormLinkDto } from "~/types";
 
 type NewLinkMutationArgs = {
