+++ apps/studio/src/utils/hooks/queries/useThemes.ts
@@ -1,6 +1,6 @@
 import { useQuery } from "react-query";
 import { getThemes } from "~/services/supabase";
-import type { Theme } from "~/types";
+import type { Theme } from "@biolnk/core";
 
 export default function useThemes() {
   const { data, ...methods } = useQuery<Theme[]>("themes", () => getThemes(), {
